Separate test_data/_dummy_pkg
