+++ pkg/pub_integration/test_data/_dummy_pkg/.gitignore
@@ -1,0 +1,2 @@
+pubspec.lock
+example/pubspec.lock
