+++ pkg/pub_integration/test_data/_dummy_pkg/README.md
@@ -1,0 +1,7 @@
+Dummy Package
+=============
+
+**Please ignore this package**.
+
+This package is published from integration tests for `pub.dev` to ensure that
+publishing keeps working, please ignore it.
