+++ pkg/pub_integration/test_data/_dummy_pkg/example/bin/main.dart
@@ -1,0 +1,4 @@
+// example
+import 'package:_dummy_pkg/_dummy_pkg.dart';
+
+void main() => sayHello();
