+++ pkg/pub_integration/test_data/_dummy_pkg/example/pubspec.yaml
@@ -1,0 +1,6 @@
+name: _dummy_pkg_example
+publish_to: none
+dependencies:
+  _dummy_pkg: ^0.0.1
+environment:
+  sdk: ">=2.1.0 <3.0.0"
