+++ pkg/pub_integration/test_data/_dummy_pkg/lib/_dummy_pkg.dart
@@ -1,0 +1,8 @@
+import 'package:retry/retry.dart';
+
+/// Say hello
+void sayHello() {
+  retry(() {
+    print('hello world');
+  }, retryIf: (e) => true); // retry, if there was an Exception
+}
