+++ pkg/pub_integration/test_data/_dummy_pkg/pubspec.yaml
@@ -1,0 +1,8 @@
+name: _dummy_pkg
+author: 'Developer <developer@example.com>'
+homepage: 'https://github.com/dart-lang/pub-dev'
+description: 'Dummy packages published from pub integration tests.'
+dependencies:
+  retry: ^2.0.0
+environment:
+  sdk: ">=2.1.0 <3.0.0"
